% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachematrix.R
\name{cacheSolve}
\alias{cacheSolve}
\title{Cache Solve}
\usage{
cacheSolve(x, ...)
}
\arguments{
\item{x}{A matrix}

\item{...}{additional arguments to be passed to \code{\link{solve}}}
}
\value{
Returns a matrix that is the inverse of \code{x}
}
\description{
This function computes the inverse of the special
  \code{\link{matrix}} returned by \code{\link{makeCacheMatrix}}. If the
  inverse has already been calculated (and the matrix has not changed), then
  the \code{cacheSolve} retrieves the inverse from the cache.
}
\examples{
x <- makeCacheMatrix(matrix(rnorm(9), 3, 3))
cacheSolve(x)
cacheSolve(x)
}
