% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachematrix.R
\name{makeCacheMatrix}
\alias{makeCacheMatrix}
\title{Make Cache Matrix}
\usage{
makeCacheMatrix(x)
}
\arguments{
\item{x}{A matrix (optional)}
}
\value{
The list functions to \code{set} and \code{get} the cached
  \code{\link{matrix}} and the its inverse (\code{setInv} and \code{getInv}).
}
\description{
This function creates a special \code{\link{matrix}} object that
  can cache its inverse.
}
\examples{
makeCacheMatrix()
makeCacheMatrix(matrix(rnorm(9), 3, 3))
}
